+++ Compare_Latest/PrivateFrameworks/HomeSharing/HSArtworkRequest.h
@@ -10,9 +10,9 @@
 {
 }
 
-+ (id)requestWithDatabaseID:(unsigned int)arg1 itemID:(unsigned int)arg2 size:(struct CGSize)arg3;
++ (id)requestWithDatabaseID:(unsigned int)arg1 itemID:(unsigned long long)arg2 size:(struct CGSize)arg3;
 - (BOOL)acceptsGzipEncoding;
-- (id)initWithDatabaseID:(unsigned int)arg1 itemID:(unsigned int)arg2 size:(struct CGSize)arg3;
+- (id)initWithDatabaseID:(unsigned int)arg1 itemID:(unsigned long long)arg2 size:(struct CGSize)arg3;
 
 @end
