+++ Compare_Latest/PrivateFrameworks/OfficeImport/CPEnumerationMap.h
@@ -14,7 +14,7 @@
 }
 
 - (id)stringForValue:(int)arg1;
-- (int)valueForString:(id)arg1;
+- (long)valueForString:(id)arg1;
 - (void)dealloc;
 - (id)initWithStructs:(const struct CPEnumerationStruct *)arg1 count:(int)arg2 caseSensitive:(BOOL)arg3;
 - (id)initWithStructs:(const struct CPEnumerationStruct *)arg1 count:(int)arg2;
