+++ Compare_Latest/PrivateFrameworks/OfficeImport/EBGraphic.h
@@ -16,7 +16,7 @@
 + (int)objectTypeForShape:(id)arg1;
 + (void)readNotesWithDictionary:(id)arg1 state:(id)arg2;
 + (void)readImage:(id)arg1 xlGraphicsInfo:(struct XlGraphicsInfo *)arg2 state:(id)arg3;
-+ (void)readChart:(id)arg1 chartIndex:(int)arg2 state:(id)arg3;
++ (void)readChart:(id)arg1 chartIndex:(short)arg2 state:(id)arg3;
 + (void)readOle:(id)arg1 xlGraphicsInfo:(struct XlGraphicsInfo *)arg2 state:(id)arg3;
 + (id)readGraphicWithDictionary:(id)arg1 state:(id)arg2;
