+++ Compare_Latest/PrivateFrameworks/OfficeImport/EDPivotConditionalFormat.h
@@ -10,7 +10,7 @@
 
 @interface EDPivotConditionalFormat : NSObject
 {
-    unsigned int mPriority;
+    unsigned long mPriority;
     int mType;
     int mScope;
     EDCollection *mPivotAreas;
@@ -22,8 +22,8 @@
 - (int)scope;
 - (void)setType:(int)arg1;
 - (int)type;
-- (void)setPriority:(unsigned int)arg1;
-- (unsigned int)priority;
+- (void)setPriority:(unsigned long)arg1;
+- (unsigned long)priority;
 - (void)dealloc;
 - (id)init;
