+++ Compare_Latest/PrivateFrameworks/OfficeImport/MFFont.h
@@ -12,11 +12,11 @@
 
 @interface MFFont : NSObject <MFObject>
 {
-    long m_lfHeight;
-    long m_lfWidth;
-    long m_lfEscapement;
-    long m_lfOrientation;
-    long m_lfWeight;
+    int m_lfHeight;
+    int m_lfWidth;
+    int m_lfEscapement;
+    int m_lfOrientation;
+    int m_lfWeight;
     _Bool m_lfItalic;
     _Bool m_lfUnderline;
     _Bool m_lfStrikeOut;
@@ -37,8 +37,8 @@
 }
 
 + (id)fontWithStockFont:(int)arg1;
-+ (id)fontWithExtendedFeatures:(long)arg1:(long)arg2:(long)arg3:(long)arg4:(long)arg5:(_Bool)arg6:(_Bool)arg7:(_Bool)arg8:(int)arg9:(int)arg10:(unsigned char)arg11:(int)arg12:(int)arg13:(int)arg14:(id)arg15:(id)arg16:(id)arg17:(unsigned int)arg18:(unsigned int)arg19:(unsigned int)arg20:(unsigned int)arg21:(unsigned int)arg22;
-+ (id)fontWithFeatures:(long)arg1:(long)arg2:(long)arg3:(long)arg4:(long)arg5:(_Bool)arg6:(_Bool)arg7:(_Bool)arg8:(int)arg9:(int)arg10:(unsigned char)arg11:(int)arg12:(int)arg13:(int)arg14:(id)arg15;
++ (id)fontWithExtendedFeatures:(int)arg1:(int)arg2:(int)arg3:(int)arg4:(int)arg5:(_Bool)arg6:(_Bool)arg7:(_Bool)arg8:(int)arg9:(int)arg10:(unsigned char)arg11:(int)arg12:(int)arg13:(int)arg14:(id)arg15:(id)arg16:(id)arg17:(unsigned int)arg18:(unsigned int)arg19:(unsigned int)arg20:(unsigned int)arg21:(unsigned int)arg22;
++ (id)fontWithFeatures:(int)arg1:(int)arg2:(int)arg3:(int)arg4:(int)arg5:(_Bool)arg6:(_Bool)arg7:(_Bool)arg8:(int)arg9:(int)arg10:(unsigned char)arg11:(int)arg12:(int)arg13:(int)arg14:(id)arg15;
 - (int)selectInto:(id)arg1;
 - (struct CGRect)textExtent:(id)arg1:(int)arg2:(int *)arg3:(int)arg4;
 - (id)faceName;
@@ -49,8 +49,8 @@
 - (double)escapement;
 - (void)dealloc;
 - (id)initWithStockFont:(int)arg1;
-- (id)initWithFeatures:(long)arg1:(long)arg2:(long)arg3:(long)arg4:(long)arg5:(_Bool)arg6:(_Bool)arg7:(_Bool)arg8:(int)arg9:(int)arg10:(unsigned char)arg11:(int)arg12:(int)arg13:(int)arg14:(id)arg15;
-- (id)initWithAllFeatures:(long)arg1:(long)arg2:(long)arg3:(long)arg4:(long)arg5:(_Bool)arg6:(_Bool)arg7:(_Bool)arg8:(int)arg9:(int)arg10:(unsigned char)arg11:(int)arg12:(int)arg13:(int)arg14:(id)arg15:(id)arg16:(id)arg17:(unsigned int)arg18:(unsigned int)arg19:(unsigned int)arg20:(unsigned int)arg21:(unsigned int)arg22;
+- (id)initWithFeatures:(int)arg1:(int)arg2:(int)arg3:(int)arg4:(int)arg5:(_Bool)arg6:(_Bool)arg7:(_Bool)arg8:(int)arg9:(int)arg10:(unsigned char)arg11:(int)arg12:(int)arg13:(int)arg14:(id)arg15;
+- (id)initWithAllFeatures:(int)arg1:(int)arg2:(int)arg3:(int)arg4:(int)arg5:(_Bool)arg6:(_Bool)arg7:(_Bool)arg8:(int)arg9:(int)arg10:(unsigned char)arg11:(int)arg12:(int)arg13:(int)arg14:(id)arg15:(id)arg16:(id)arg17:(unsigned int)arg18:(unsigned int)arg19:(unsigned int)arg20:(unsigned int)arg21:(unsigned int)arg22;
 - (id)stringWithBytes:(const void *)arg1 length:(unsigned int)arg2;
 
 @end
